cjwkernel: introduce the idea of "framework"
Now, we can convert thrift<=>pandas directly instead of going
thrift<=>arrow<=>pandas. The double-conversion idea duplicated all our
bad decisions. For example: Arrow modules got the
"empty-fetched-file-is-empty-parquet-dataframe" feature ... even though
we don't want that feature.)

With each commit, I'm reducing the number of conversion functions and
hops. The big prize is Params....
