+++ cjwkernel/tests/pandas/test_types.py
@@ -7,8 +7,8 @@ from typing import List
 
 import numpy as np
 import pandas as pd
-import pyarrow
-from pandas.testing import assert_frame_equal, assert_series_equal
+import pyarrow as pa
+from pandas.testing import assert_frame_equal
 from cjwmodule.arrow.testing import make_table, make_column, assert_arrow_table_equals
 
 import cjwkernel.types as atypes
@@ -17,14 +17,15 @@ from cjwkernel.pandas.types import (
     Column,
     ColumnType,
     ProcessResult,
+    RenderColumn,
     RenderError,
     QuickFix,
     QuickFixAction,
-    arrow_table_to_dataframe,
     coerce_I18nMessage,
     coerce_RenderError_list,
     coerce_RenderError,
     dataframe_to_arrow_table,
+    arrow_schema_to_render_columns,
 )
 from cjwkernel.tests.util import override_settings, tempfile_context
 from cjwkernel.util import create_tempfile
@@ -976,7 +977,6 @@ class ProcessResultTests(unittest.TestCase):
         os.write(fd, b"to-remove")
         os.close(fd)
         try:
-
             result = ProcessResult.coerce("bad, bad error").to_arrow(Path(filename))
             self.assertEqual(
                 result,
@@ -985,12 +985,11 @@ class ProcessResultTests(unittest.TestCase):
                     {},
                 ),
             )
-            self.assertEqual(Path(filename).stat().st_size, 0)
+            assert_arrow_table_equals(
+                load_untrusted_arrow_file_with_columns(Path(filename))[0], make_table()
+            )
         finally:
-            try:
-                os.unlink(filename)
-            except FileNotFoundError:
-                pass
+            os.unlink(filename)
 
     def test_to_arrow_normal_dataframe(self):
         fd, filename = tempfile.mkstemp()
@@ -1004,7 +1003,7 @@ class ProcessResultTests(unittest.TestCase):
                     {},
                 ),
             )
-            with pyarrow.ipc.open_file(filename) as reader:
+            with pa.ipc.open_file(filename) as reader:
                 table = reader.read_all()
             assert_arrow_table_equals(
                 table,
@@ -1034,7 +1033,7 @@ class ArrowConversionTests(unittest.TestCase):
         self,
         dataframe: pd.DataFrame,
         columns: List[Column],
-        expected_table: pyarrow.Table,
+        expected_table: pa.Table,
     ) -> None:
         with tempfile_context() as path:
             dataframe_to_arrow_table(dataframe, columns, path)
@@ -1047,66 +1046,49 @@ class ArrowConversionTests(unittest.TestCase):
         self._test_dataframe_to_arrow_table(
             pd.DataFrame({"A": [None]}, dtype=str),
             [Column("A", ColumnType.Text())],
-            expected_table=make_table(make_column("A", [None], pyarrow.string())),
+            expected_table=make_table(make_column("A", [None], pa.string())),
         )
 
-    def test_arrow_all_null_text_column(self):
-        dataframe, columns = arrow_table_to_dataframe(
-            make_table(make_column("A", [None], pyarrow.string()))
+    def test_arrow_schema_text_column(self):
+        self.assertEqual(
+            arrow_schema_to_render_columns(pa.schema([pa.field("A", pa.string())])),
+            {"A": RenderColumn("A", "text", None)},
         )
-        assert_frame_equal(dataframe, pd.DataFrame({"A": [None]}, dtype=str))
-        self.assertEqual(columns, [Column("A", ColumnType.Text())])
 
     def test_dataframe_category_column(self):
         self._test_dataframe_to_arrow_table(
             pd.DataFrame({"A": ["A", "B", None, "A"]}, dtype="category"),
             [Column("A", ColumnType.Text())],
-            pyarrow.table(
+            pa.table(
                 {
-                    "A": pyarrow.DictionaryArray.from_arrays(
-                        pyarrow.array([0, 1, None, 0], pyarrow.int8()),
-                        pyarrow.array(["A", "B"], pyarrow.string()),
+                    "A": pa.DictionaryArray.from_arrays(
+                        pa.array([0, 1, None, 0], pa.int8()),
+                        pa.array(["A", "B"], pa.string()),
                     ),
                 }
             ),
         )
 
-    def test_arrow_category_column(self):
-        dataframe, columns = arrow_table_to_dataframe(
-            make_table(make_column("A", ["A", "B", None, "A"], dictionary=True))
-        )
-        self.assertEqual(columns, [Column("A", ColumnType.Text())])
-        assert_frame_equal(
-            dataframe, pd.DataFrame({"A": ["A", "B", None, "A"]}, dtype="category")
+    def test_arrow_schema_category_column(self):
+        self.assertEqual(
+            arrow_schema_to_render_columns(
+                pa.schema([pa.field("A", pa.dictionary(pa.int32(), pa.string()))])
+            ),
+            {"A": RenderColumn("A", "text", None)},
         )
 
     def test_dataframe_all_null_category_column(self):
         self._test_dataframe_to_arrow_table(
             pd.DataFrame({"A": [None]}, dtype=str).astype("category"),
             [Column("A", ColumnType.Text())],
-            pyarrow.table(
+            pa.table(
                 {
-                    "A": pyarrow.DictionaryArray.from_arrays(
-                        pyarrow.array([None], pyarrow.int8()),
-                        pyarrow.array([], pyarrow.string()),
+                    "A": pa.DictionaryArray.from_arrays(
+                        pa.array([None], pa.int8()),
+                        pa.array([], pa.string()),
                     ),
                 }
             ),
-        )
-
-    def test_arrow_all_null_category_column(self):
-        atable = pyarrow.table(
-            {
-                "A": pyarrow.DictionaryArray.from_arrays(
-                    pyarrow.array([None], pyarrow.int8()),
-                    pyarrow.array([], pyarrow.string()),
-                ),
-            }
-        )
-        dataframe, columns = arrow_table_to_dataframe(atable)
-        self.assertEqual(columns, [Column("A", ColumnType.Text())])
-        assert_frame_equal(
-            dataframe, pd.DataFrame({"A": [None]}, dtype=str).astype("category")
         )
 
     def test_dataframe_uint8_column(self):
@@ -1114,20 +1096,17 @@ class ArrowConversionTests(unittest.TestCase):
             pd.DataFrame({"A": [1, 2, 3, 253]}, dtype=np.uint8),
             [Column("A", ColumnType.Number("{:,d}"))],
             make_table(
-                make_column("A", [1, 2, 3, 253], type=pyarrow.uint8(), format="{:,d}")
+                make_column("A", [1, 2, 3, 253], type=pa.uint8(), format="{:,d}")
             ),
         )
 
-    def test_arrow_uint8_column(self):
-        dataframe, columns = arrow_table_to_dataframe(
-            make_table(
-                make_column("A", [1, 2, 3, 253], type=pyarrow.uint8(), format="{:,d}")
+    def test_arrow_schema_uint8_column(self):
+        self.assertEqual(
+            arrow_schema_to_render_columns(
+                pa.schema([pa.field("A", pa.uint8(), metadata={"format": "{:,d}"})])
             ),
+            {"A": RenderColumn("A", "number", "{:,d}")},
         )
-        assert_frame_equal(
-            dataframe, pd.DataFrame({"A": [1, 2, 3, 253]}, dtype=np.uint8)
-        )
-        self.assertEqual(columns, [Column("A", ColumnType.Number("{:,d}"))])
 
     def test_dataframe_datetime_column(self):
         self._test_dataframe_to_arrow_table(
@@ -1141,21 +1120,17 @@ class ArrowConversionTests(unittest.TestCase):
         )
 
     def test_arrow_timestamp_column(self):
-        dataframe, columns = arrow_table_to_dataframe(
-            make_table(
-                make_column("A", [dt.fromisoformat("2019-09-17T21:21:00.123456"), None])
+        self.assertEqual(
+            arrow_schema_to_render_columns(
+                pa.schema([pa.field("A", pa.timestamp("ns"))])
             ),
+            {"A": RenderColumn("A", "timestamp", None)},
         )
-        assert_frame_equal(
-            dataframe,
-            pd.DataFrame(
-                {"A": ["2019-09-17T21:21:00.123456Z", None]}, dtype="datetime64[ns]"
-            ),
-        )
-        self.assertEqual(columns, [Column("A", ColumnType.Timestamp())])
 
-    def test_arrow_table_reuse_string_memory(self):
-        dataframe, _ = arrow_table_to_dataframe(
-            make_table(make_column("A", ["x", "x"]))
+    def test_arrow_date32_column(self):
+        self.assertEqual(
+            arrow_schema_to_render_columns(
+                pa.schema([pa.field("A", pa.date32(), metadata={"unit": "month"})])
+            ),
+            {"A": RenderColumn("A", "date", "month")},
         )
-        self.assertIs(dataframe["A"][0], dataframe["A"][1])
